refactor: restructure "pc import" to support multiple import types and improve command organization
